requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[blacklist]
django-cors-headers
django-filter
pillow
python-dotenv
python-decouple
requests
stripe
weasyprint
mysqlclient